[refactor] reset api and domain module
