+++ api/src/main/java/clov3r/api/friend/controller/FriendController.java
@@ -3,8 +3,8 @@ package clov3r.api.friend.controller;
 import clov3r.api.auth.service.UserService;
 import clov3r.api.friend.domain.dto.OtherUserDTO;
 import clov3r.api.friend.domain.dto.RequestFriendDTO;
-import clov3r.domain.error.errorcode.CustomErrorCode;
-import clov3r.domain.error.exception.BaseExceptionV2;
+import clov3r.api.common.error.errorcode.CustomErrorCode;
+import clov3r.api.common.error.exception.BaseExceptionV2;
 import clov3r.api.auth.security.Auth;
 import clov3r.api.friend.domain.dto.FriendDTO;
 import clov3r.api.friend.domain.dto.FriendReqDTO;
