+++ api/src/main/java/clov3r/api/giftbox/controller/GiftboxProductController.java
@@ -1,12 +1,12 @@
 package clov3r.api.giftbox.controller;
 
-import static clov3r.domain.error.errorcode.CommonErrorCode.REQUEST_ERROR;
-import static clov3r.domain.error.errorcode.CustomErrorCode.GIFTBOX_NOT_FOUND;
-import static clov3r.domain.error.errorcode.CustomErrorCode.GIFTBOX_PRODUCT_NOT_FOUND;
-import static clov3r.domain.error.errorcode.CustomErrorCode.NOT_PARTICIPANT_OF_GIFTBOX;
-import static clov3r.domain.error.errorcode.CustomErrorCode.PRODUCT_NOT_FOUND;
-import static clov3r.domain.error.errorcode.CustomErrorCode.SEARCH_KEYWORD_ERROR;
-import static clov3r.domain.error.errorcode.CustomErrorCode.USER_NOT_FOUND;
+import static clov3r.api.common.error.errorcode.CommonErrorCode.REQUEST_ERROR;
+import static clov3r.api.common.error.errorcode.CustomErrorCode.GIFTBOX_NOT_FOUND;
+import static clov3r.api.common.error.errorcode.CustomErrorCode.GIFTBOX_PRODUCT_NOT_FOUND;
+import static clov3r.api.common.error.errorcode.CustomErrorCode.NOT_PARTICIPANT_OF_GIFTBOX;
+import static clov3r.api.common.error.errorcode.CustomErrorCode.PRODUCT_NOT_FOUND;
+import static clov3r.api.common.error.errorcode.CustomErrorCode.SEARCH_KEYWORD_ERROR;
+import static clov3r.api.common.error.errorcode.CustomErrorCode.USER_NOT_FOUND;
 
 import clov3r.api.auth.security.Auth;
 import clov3r.api.giftbox.repository.Giftbox.GiftboxRepository;
@@ -18,7 +18,7 @@ import clov3r.domain.domains.entity.Product;
 import clov3r.domain.domains.entity.id.GiftboxProductVoteId;
 import clov3r.api.giftbox.domain.dto.GiftboxProductDetailDTO;
 import clov3r.api.giftbox.domain.request.VoteProductRequest;
-import clov3r.domain.error.exception.BaseExceptionV2;
+import clov3r.api.common.error.exception.BaseExceptionV2;
 import clov3r.api.giftbox.repository.GiftboxProduct.GiftboxProductRepository;
 import clov3r.api.product.repository.ProductRepository;
 import clov3r.api.auth.repository.UserRepository;
