+++ api/src/main/java/clov3r/api/product/controller/ProductControllerV2.java
@@ -1,9 +1,9 @@
 package clov3r.api.product.controller;
 
 import clov3r.api.product.domain.dto.ProductDTO;
-import clov3r.domain.error.errorcode.CommonErrorCode;
-import clov3r.domain.error.errorcode.CustomErrorCode;
-import clov3r.domain.error.exception.BaseExceptionV2;
+import clov3r.api.common.error.errorcode.CommonErrorCode;
+import clov3r.api.common.error.errorcode.CustomErrorCode;
+import clov3r.api.common.error.exception.BaseExceptionV2;
 import clov3r.api.auth.security.Auth;
 import clov3r.api.product.domain.dto.ProductDetailDTO;
 import clov3r.api.product.domain.dto.ProductSummaryDTO;
